Update  - Enhance return error in SPO Engine  - Split download/upload to new module
